Add files via upload
Threshold added to ML procedure so that instead of picking the history length with the lowest BIC score, the smallest history length within some threshold of the lowest BIC score is chosen (set by user). This prevents tiny differences in BIC that are probably uninformative from inflating the history length.